+++ src/command_result.rs
@@ -24,10 +24,11 @@ impl CommandResult {
         output: CommandOutput,
         exit_status: Option<ExitStatus>,
         analyzer: Analyzer,
+        mission: &Mission,
     ) -> Result<Self> {
         let lines = &output.lines;
         let error_code = exit_status.and_then(|s| s.code()).filter(|&c| c != 0);
-        let mut report = analyzer.build_report(lines)?;
+        let mut report = analyzer.build_report(lines, mission)?;
         debug!("report stats: {:?}", &report.stats);
         if let Some(error_code) = error_code {
             if report.stats.errors + report.stats.test_fails == 0 {